+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/magical/MagicalZombiePills.java
@@ -85,6 +85,7 @@ public class MagicalZombiePills extends SimpleSlimefunItem<EntityInteractHandler
             ItemUtils.consumeItem(item, false);
         }
 
+        // This is supposed to be a vanilla sound. No need for a SoundEffect
         p.playSound(p.getLocation(), Sound.ENTITY_ZOMBIE_VILLAGER_CONVERTED, 1, 1);
     }