+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 * Nether Wart Blocks can now be turned into Nether Warts using a Grind Stone
 * Added an option to allow Talismans to send their notifications via the Actionbar
 * /sf versions now shows the Java version and some useful tooltips
+* Added sound effects for Slime Boots
+* Added sounds.yml file to configure all sound effects
 
 #### Changes
 * Deprecated Automatic Crafting Chamber